refactor: 카드 조회 및 수정 코드 정리
